+++ dlls/win32u/hook.c
@@ -129,7 +129,7 @@ HHOOK WINAPI NtUserSetWindowsHookEx( HINSTANCE inst, UNICODE_STRING *module, DWO
     }
     SERVER_END_REQ;
 
-    TRACE( "%s %p %x -> %p\n", debugstr_hook_id(id), proc, tid, handle );
+    TRACE( "%s %p %x -> %p\n", debugstr_hook_id(id), proc, (int)tid, handle );
     return handle;
 }
 
@@ -206,7 +206,8 @@ static LRESULT call_hook( struct win_hook_params *info, const WCHAR *module )
         h_extra.lparam = info->lparam;
 
         TRACE( "calling hook in thread %04x %s code %x wp %lx lp %lx\n",
-               info->tid, hook_names[info->id-WH_MINHOOK], info->code, info->wparam, info->lparam );
+               (int)info->tid, hook_names[info->id-WH_MINHOOK],
+               info->code, (long)info->wparam, info->lparam );
 
         switch(info->id)
         {
@@ -307,7 +308,7 @@ static LRESULT call_hook( struct win_hook_params *info, const WCHAR *module )
         }
 
         TRACE( "calling hook %p %s code %x wp %lx lp %lx module %s\n",
-               params->proc, hook_names[params->id-WH_MINHOOK], params->code, params->wparam,
+               params->proc, hook_names[params->id-WH_MINHOOK], params->code, (long)params->wparam,
                params->lparam, debugstr_w(module) );
 
         thread_info->hook = params->handle;
@@ -532,7 +533,7 @@ void WINAPI NtUserNotifyWinEvent( DWORD event, HWND hwnd, LONG object_id, LONG c
     ULONG ret_len;
     BOOL ret;
 
-    TRACE( "%04x, %p, %d, %d\n", event, hwnd, object_id, child_id );
+    TRACE( "%04x, %p, %d, %d\n", (int)event, hwnd, (int)object_id, (int)child_id );
 
     user_check_not_lock();
 
@@ -577,7 +578,7 @@ void WINAPI NtUserNotifyWinEvent( DWORD event, HWND hwnd, LONG object_id, LONG c
     do
     {
         TRACE( "calling WH_WINEVENT hook %p event %x hwnd %p %x %x module %s\n",
-               info.proc, event, hwnd, object_id, child_id, debugstr_w(info.module) );
+               info.proc, (int)event, hwnd, (int)object_id, (int)child_id, debugstr_w(info.module) );
 
         info.time = NtGetTickCount();
         KeUserModeCallback( NtUserCallWinEventHook, &info,
