+++ dlls/win32u/painting.c
@@ -910,7 +910,7 @@ BOOL WINAPI NtGdiGradientFill( HDC hdc, TRIVERTEX *vert_array, ULONG nvert,
  */
 BOOL WINAPI NtGdiDrawStream( HDC hdc, ULONG in, void *pvin )
 {
-    FIXME("stub: %p, %d, %p\n", hdc, in, pvin);
+    FIXME("stub: %p, %d, %p\n", hdc, (int)in, pvin);
     return FALSE;
 }
