+++ dlls/win32u/scroll.c
@@ -663,7 +663,7 @@ void handle_scroll_event( HWND hwnd, int bar, UINT msg, POINT pt )
     }
 
     TRACE( "Event: hwnd=%p bar=%d msg=%s pt=%d,%d hit=%d\n",
-           hwnd, bar, debugstr_msg_name( msg, hwnd ), pt.x, pt.y, hittest );
+           hwnd, bar, debugstr_msg_name( msg, hwnd ), (int)pt.x, (int)pt.y, hittest );
 
     switch (g_tracking_info.hit_test)
     {
@@ -1188,7 +1188,7 @@ static void create_scroll_bar( HWND hwnd, CREATESTRUCTW *create )
 
 static void handle_kbd_event( HWND hwnd, WPARAM wparam, LPARAM lparam )
 {
-    TRACE( "hwnd=%p wparam=%ld lparam=%ld\n", hwnd, wparam, lparam );
+    TRACE( "hwnd=%p wparam=%ld lparam=%ld\n", hwnd, (long)wparam, lparam );
 
     /* hide caret on first KEYDOWN to prevent flicker */
     if ((lparam & PFD_DOUBLEBUFFER_DONTCARE) == 0)
@@ -1420,12 +1420,12 @@ LRESULT scroll_bar_window_proc( HWND hwnd, UINT msg, WPARAM wparam, LPARAM lpara
     case 0x00ed:
     case 0x00ee:
     case 0x00ef:
-        ERR( "unknown Win32 msg %04x wp=%08lx lp=%08lx\n", msg, wparam, lparam );
+        ERR( "unknown Win32 msg %04x wp=%08lx lp=%08lx\n", msg, (long)wparam, lparam );
         return 0;
 
     default:
         if (msg >= WM_USER)
-            WARN( "unknown msg %04x wp=%08lx lp=%08lx\n", msg, wparam, lparam );
+            WARN( "unknown msg %04x wp=%08lx lp=%08lx\n", msg, (long)wparam, lparam );
         return default_window_proc( hwnd, msg, wparam, lparam, ansi );
     }
 }
@@ -1460,7 +1460,7 @@ BOOL WINAPI NtUserShowScrollBar( HWND hwnd, INT bar, BOOL show )
  */
 BOOL WINAPI NtUserGetScrollBarInfo( HWND hwnd, LONG id, SCROLLBARINFO *info )
 {
-    TRACE( "hwnd=%p id=%d info=%p\n", hwnd, id, info );
+    TRACE( "hwnd=%p id=%d info=%p\n", hwnd, (int)id, info );
 
     /* Refer OBJID_CLIENT requests to the window */
     if (id == OBJID_CLIENT)
