+++ dlls/win32u/winstation.c
@@ -253,10 +253,10 @@ HDESK WINAPI NtUserOpenInputDesktop( DWORD flags, BOOL inherit, ACCESS_MASK acce
 {
     HANDLE ret = 0;
 
-    TRACE( "(%x,%i,%x)\n", flags, inherit, access );
+    TRACE( "(%x,%i,%x)\n", (int)flags, inherit, (int)access );
 
     if (flags)
-        FIXME( "partial stub flags %08x\n", flags );
+        FIXME( "partial stub flags %08x\n", (int)flags );
 
     SERVER_START_REQ( open_input_desktop )
     {
@@ -513,7 +513,7 @@ static HANDLE get_winstations_dir_handle(void)
     NTSTATUS status;
     HANDLE dir;
 
-    sprintf( bufferA, "\\Sessions\\%u\\Windows\\WindowStations", NtCurrentTeb()->Peb->SessionId );
+    sprintf( bufferA, "\\Sessions\\%u\\Windows\\WindowStations", (int)NtCurrentTeb()->Peb->SessionId );
     str.Buffer = buffer;
     str.Length = str.MaximumLength = asciiz_to_unicode( buffer, bufferA ) - sizeof(WCHAR);
     InitializeObjectAttributes( &attr, &str, 0, 0, NULL );
